Implement count_archives
git-svn-id: https://svn.macosforge.org/repository/darwinbuild/branches/PR-7489777@709 10a61168-4876-4dac-953b-31e694342555

+++ darwinup/Depot.h
@@ -72,8 +72,8 @@ struct Depot {
 	Archive* get_archive(const char* arg);
 
 	// returns a list of Archive*. Caller must free the list. 
-	Archive** get_all_archives(size_t *count);
-	size_t count_archives();
+	Archive** get_all_archives(uint64_t *count);
+	uint64_t count_archives();
 	
 	int dump();
 	static int dump_archive(Archive* archive, void* context);
